refactor: change test file names
