+++ internal/sbomscannerdb/oci/manifest_test.go
